user null problem solved

+++ frontend/src/features/goals/goalSlice.js
@@ -14,8 +14,10 @@ export const createGoal = createAsyncThunk(
   "goals/create",
   async (goalData, thunkAPI) => {
     try {
-      const token = thunkAPI.getState().auth.user.token;
-      return await goalService.createGoal(goalData, token);
+      if (thunkAPI.getState().auth.user) {
+        const token = thunkAPI.getState().auth.user.token;
+        return await goalService.createGoal(goalData, token);
+      }
     } catch (error) {
       const message =
         (error.response &&
@@ -34,8 +36,10 @@ export const getGoals = createAsyncThunk(
   "goals/getAll",
   async (_, thunkAPI) => {
     try {
-      const token = thunkAPI.getState().auth.user.token;
-      return await goalService.getGoals(token);
+      if (thunkAPI.getState().auth.user) {
+        const token = thunkAPI.getState().auth.user.token;
+        return await goalService.getGoals(token);
+      }
     } catch (error) {
       const message =
         (error.response &&
@@ -54,8 +58,10 @@ export const deleteGoal = createAsyncThunk(
   "goals/delete",
   async (id, thunkAPI) => {
     try {
-      const token = thunkAPI.getState().auth.user.token;
-      return await goalService.deleteGoal(id, token);
+      if (thunkAPI.getState().auth.user) {
+        const token = thunkAPI.getState().auth.user.token;
+        return await goalService.deleteGoal(id, token);
+      }
     } catch (error) {
       const message =
         (error.response &&
@@ -108,13 +114,15 @@ export const goalSlice = createSlice({
       .addCase(deleteGoal.fulfilled, (state, action) => {
         state.isLoading = false;
         state.isSuccess = true;
-        state.goals = state.goals.filter((goal) => goal._id !== action.payload.id);
+        state.goals = state.goals.filter(
+          (goal) => goal._id !== action.payload.id
+        );
       })
       .addCase(deleteGoal.rejected, (state, action) => {
         state.isLoading = false;
         state.isError = true;
         state.message = action.payload;
-      })
+      });
   },
 });
